pyproject.toml
[project]
name = "setup-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.10.5",
    "python-dotenv>=1.1.1",
]
